refactor(context): reorganize the context module of the integration package
