--- packages/integration/src/lib/context-manager/context-editor-tool/index.ts
@@ -1,1 +1,0 @@
-export * from './context-editor-tool.component';
