--- packages/integration/src/lib/context-manager/context-manager-tool/index.ts
@@ -1,1 +1,0 @@
-export * from './context-manager-tool.component';
