--- packages/integration/src/lib/context-manager/index.ts
@@ -1,4 +1,0 @@
-export * from './context-manager-tool';
-export * from './context-editor-tool';
-export * from './permissions-context-manager-tool';
-export * from './context-manager.module';
