--- packages/integration/src/lib/context-manager/permissions-context-manager-tool/index.ts
@@ -1,1 +1,0 @@
-export * from './permissions-context-manager-tool.component';
