--- packages/integration/src/lib/context-manager/tools-context-manager-tool/index.ts
@@ -1,1 +1,0 @@
-export * from './tools-context-manager-tool.component';
