--- packages/integration/src/lib/context-manager/tools-context-manager-tool/tools-context-manager-tool.component.html
@@ -1,1 +1,0 @@
-<igo-context-tools igoContextToolsBinding></igo-context-tools>
