+++ packages/integration/src/lib/context/context-permission-manager-tool/context-permission-manager-tool.component.html
