+++ packages/integration/src/lib/context/context-permission-manager-tool/context-permission-manager-tool.component.ts
@@ -3,14 +3,14 @@ import { Component, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { ToolComponent } from '@igo2/common';
 
 @ToolComponent({
-  name: 'toolsContextManager',
+  name: 'contextPermissionManager',
   title: 'igo.integration.tools.contexts',
   icon: 'settings'
 })
 @Component({
-  selector: 'igo-tools-context-manager-tool',
-  templateUrl: './tools-context-manager-tool.component.html'
+  selector: 'igo-context-permission-manager-tool',
+  templateUrl: './context-permission-manager-tool.component.html'
 })
-export class ToolsContextManagerToolComponent {
+export class ContextPermissionManagerToolComponent {
   constructor() {}
 }
