+++ packages/integration/src/lib/integration.module.ts
@@ -1,13 +1,12 @@
 import { NgModule } from '@angular/core';
 
-import { IgoAppContextManagerModule } from './context-manager/context-manager.module';
+import { IgoAppContextModule } from './context/context.module';
 import { IgoAppCatalogModule } from './catalog/catalog.module';
 import { IgoAppDirectionsModule } from './directions/directions.module';
 import { IgoAppImportExportModule } from './import-export/import-export.module';
 import { IgoAppMapModule } from './map/map.module';
 import { IgoAppPrintModule } from './print/print.module';
 import { IgoAppSearchModule } from './search/search.module';
-import { IgoAppShareMapModule } from './share-map/share-map.module';
 import { IgoAppFilterModule } from './filter/filter.module';
 import { IgoAppAboutModule } from './about/about.module';
 
@@ -15,14 +14,13 @@ import { IgoAppAboutModule } from './about/about.module';
   imports: [],
   declarations: [],
   exports: [
-    IgoAppContextManagerModule,
+    IgoAppContextModule,
     IgoAppCatalogModule,
     IgoAppDirectionsModule,
     IgoAppImportExportModule,
     IgoAppMapModule,
     IgoAppPrintModule,
     IgoAppSearchModule,
-    IgoAppShareMapModule,
     IgoAppFilterModule,
     IgoAppAboutModule
   ]
