--- packages/integration/src/lib/share-map/index.ts
@@ -1,2 +1,0 @@
-export * from './share-map-tool';
-export * from './share-map.module';
