--- packages/integration/src/lib/share-map/share-map-tool/index.ts
@@ -1,2 +1,0 @@
-export * from './share-map-tool.component';
-export * from './share-map-tool.interface';
