--- packages/integration/src/lib/share-map/share-map-tool/share-map-tool.interface.ts
@@ -1,4 +1,0 @@
-export interface ShareMapToolOptions {
-  hasCopyLinkButton?: boolean;
-  hasShareMapButton?: boolean;
-}
