--- packages/integration/src/lib/share-map/share-map.module.ts
@@ -1,24 +1,0 @@
-import {
-  NgModule,
-  ModuleWithProviders,
-  CUSTOM_ELEMENTS_SCHEMA
-} from '@angular/core';
-
-import { IgoShareMapModule } from '@igo2/context';
-import { ShareMapToolComponent } from './share-map-tool/share-map-tool.component';
-
-@NgModule({
-  imports: [IgoShareMapModule],
-  declarations: [ShareMapToolComponent],
-  exports: [ShareMapToolComponent],
-  entryComponents: [ShareMapToolComponent],
-  schemas: [CUSTOM_ELEMENTS_SCHEMA]
-})
-export class IgoAppShareMapModule {
-  static forRoot(): ModuleWithProviders {
-    return {
-      ngModule: IgoAppShareMapModule,
-      providers: []
-    };
-  }
-}
