+++ packages/integration/src/public_api.ts
@@ -4,7 +4,7 @@
 
 export * from './lib/integration.module';
 export * from './lib/about/about.module';
-export * from './lib/context-manager/context-manager.module';
+export * from './lib/context/context.module';
 export * from './lib/catalog/catalog.module';
 export * from './lib/directions/directions.module';
 export * from './lib/filter/filter.module';
@@ -12,11 +12,10 @@ export * from './lib/import-export/import-export.module';
 export * from './lib/map/map.module';
 export * from './lib/print/print.module';
 export * from './lib/search/search.module';
-export * from './lib/share-map/share-map.module';
 export * from './lib/tool/tool.module';
 
 export * from './lib/about';
-export * from './lib/context-manager';
+export * from './lib/context';
 export * from './lib/catalog';
 export * from './lib/directions';
 export * from './lib/filter';
@@ -24,5 +23,4 @@ export * from './lib/import-export';
 export * from './lib/map';
 export * from './lib/print';
 export * from './lib/search';
-export * from './lib/share-map';
 export * from './lib/tool';
